Initialize disconnect character with number
The disconnect character was origionally being
initialized with the ascii character that is
sent whent he socket disconnects.  This ascii
character is not supported by git, which caused
issues when pulling the codebase.  We now initialize
the character with the actual integer value that is
returned instead of the character to solve this issue.